requirements.txt
streamlit>=1.36.0
pandas>=2.2.0
apache-iotdb>=1.3.0
